Code cleaned
Cleaning code
Organization of the repo
Dalamud API 10 Preparation (IPluginLogger)
