+++ Plugin/App/Configuration.cs
