+++ Plugin/App/Struct/ChannelSetting.cs
