+++ Plugin/PuppetMaster.csproj
@@ -5,11 +5,19 @@
   <PropertyGroup>
     <Authors></Authors>
     <Company></Company>
-    <Version>7.0.0.0</Version>
+    <Version>7.0.0.1</Version>
     <Description>A sample plugin.</Description>
     <Copyright></Copyright>
     <PackageProjectUrl>https://github.com/Aspher0/PuppetMaster_Fork</PackageProjectUrl>
     <PackageLicenseExpression>AGPL-3.0-or-later</PackageLicenseExpression>
     <IsPackable>false</IsPackable>
+  </PropertyGroup>
+
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
+    <DebugType>none</DebugType>
+  </PropertyGroup>
+
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
+    <DebugType>none</DebugType>
   </PropertyGroup>
 </Project>
