+++ Plugin/PuppetMaster.json
@@ -1,0 +1,17 @@
+{
+    "Author": "DodingDaga",
+    "Name": "Puppet Master",
+    "Punchline": "Let others tell you what to do.",
+    "Description": "Let people control your character through chat.",
+    "InternalName": "PuppetMaster",
+    "ApplicableVersion": "any",
+    "Tags": [
+        "puppet",
+        "master",
+        "slave",
+        "control",
+        "emote",
+        "plugin",
+        "commands"
+    ]
+}
